31071 upgrade queryPage #2
