+++ packages/common-ui/lib/index.ts
@@ -93,7 +93,6 @@ export * from "./list-page/query-builder/QueryBuilder";
 export * from "./list-page/hardcoded-queries/directMaterialSampleChildrenQuery";
 export * from "./list-page/hardcoded-queries/materialSampleCultureStrainChildrenQuery";
 export * from "./formik-connected/QueryLogicSwitchField";
-export * from "./list-page/QueryPage";
 export * from "./custom-query-view/CustomQueryPageView";
 export * from "./formik-connected/ControlledVocabularySelectField";
 export * from "./formik-connected/FieldExtensionSelectField";
@@ -109,4 +108,4 @@ export * from "./table8/RowComponents";
 export * from "./table8/stringArray-cell8";
 export * from "./table8/date-cell8";
 export * from "./table8/booleanCell8";
-export * from "./list-page/QueryPage8";
+export * from "./list-page/QueryPage";
