+++ packages/dina-ui/components/transaction/MaterialSampleTransactionList.tsx
@@ -1,9 +1,4 @@
-import {
-  useCollapsibleSection,
-  QueryPage,
-  LoadingSpinner,
-  QueryPage8
-} from "common-ui";
+import { useCollapsibleSection, LoadingSpinner, QueryPage } from "common-ui";
 import { TRANSACTION_TABLE_COLUMNS } from "../../pages/loan-transaction/transaction/list";
 
 export interface TransactionTable {
@@ -19,7 +14,7 @@ export function MaterialSampleTransactionList({ transactionQueryDSL }) {
   }
 
   return (
-    <QueryPage8
+    <QueryPage
       indexName="dina_loan_transaction_index"
       customViewElasticSearchQuery={transactionQueryDSL}
       customViewFields={[]}
